# swagger.yaml
basePath: /api
...
host: localhost:6969
info:
    description: This is api server
    title: api server
    version: "0.1"
paths:
...
    /command/{id}/undo:
        post:
            description: Undo command.
            operationId: undoCommand
            parameters:
                - example: 1
                  in: path
                  name: id
                  required: true
                  type: string
            responses: {}
            security:
                - Bearer:
                    - apiKey
            summary: Undo command.
            tags:
                - command
...
schemes:
    - http
    - https
securityDefinitions:
    Bearer:
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"